Adicionada a parte de detalhes, ainda incompleta, e componentizado todo o site
